feat(core_server): add jwt api guard
