+++ apps/core_server/src/auth/auth.module.ts
@@ -2,10 +2,12 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
+import { APP_GUARD } from '@nestjs/core';
 
 import { GoogleStrategy } from './strategies/google.strategy';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
+import { AuthGuard } from './auth.guard';
 
 import { UserModule } from '@/user/user.module';
 
@@ -21,7 +23,14 @@ import { UserModule } from '@/user/user.module';
       }),
     }),
   ],
-  providers: [GoogleStrategy, AuthService],
+  providers: [
+    GoogleStrategy,
+    AuthService,
+    {
+      provide: APP_GUARD,
+      useClass: AuthGuard,
+    },
+  ],
   controllers: [AuthController],
 })
 export class AuthModule {}
