+++ apps/core_server/src/main.ts
@@ -3,6 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
 import { json, urlencoded } from 'express';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import cookieParser from 'cookie-parser';
 
 import { AllExceptionsFilter } from './common';
 import { CoreModule } from './core.module';
@@ -26,6 +27,7 @@ async function bootstrap() {
 
   const configService = app.get(ConfigService);
 
+  app.use(cookieParser());
   app.enableCors({ origin: `http://localhost:${configService.get('NEXT_PUBLIC_PORT', 3000)}` });
 
   app.useGlobalPipes(
