+++ package.json
@@ -17,6 +17,7 @@
     "@nestjs/config": "^4.0.1",
     "@nestjs/mapped-types": "*",
     "commander": "^13.1.0",
+    "cookie-parser": "^1.4.7",
     "dotenv": "^16.4.7",
     "eslint": "^8.57.0",
     "nest-winston": "^1.10.2",
@@ -27,6 +28,7 @@
     "@stylistic/eslint-plugin": "^1.8.0",
     "@stylistic/eslint-plugin-js": "^2.1.0",
     "@stylistic/eslint-plugin-migrate": "^1.8.0",
+    "@types/cookie-parser": "^1.4.8",
     "@types/express": "^5.0.0",
     "conventional-changelog-conventionalcommits": "^4.5.0",
     "dotenv-cli": "^8.0.0",
